requirements.txt
beautifulsoup4==4.11.1
certifi==2022.12.7
charset-normalizer==2.1.1
lxml==4.9.2
requests==2.28.1
soupsieve==2.3.2.post1
urllib3==1.26.13
Django==4.1.4
django-crispy-forms==1.14.0
crispy-bootstrap5==0.7